Add ISO_Level3_Shift (0xfe03) to NimfKeySym
In some laptops, the hardware Hangul key is mapped to ISO_Level3_Shift
by default so that many Korean users have to manually remap it to
Hangul with xmodmap.

This patch adds ISO_Level3_Shift to NimfKeySym to allow users to
register it as a trigger key in Nimf settings without dealing with
xmodmap.